Remove random_fullrank_matrix_distinc_singular_value
We do so in favour of
`make_fullrank_matrices_with_distinct_singular_values` as this latter
one not only has an even longer name, but also generates inputs
correctly for them to work with the PR that tests noncontig inputs
latter in this stack.

We also heavily simplified the generation of samples for the SVD, as it was
fairly convoluted and it was not generating the inputs correclty for
the noncontiguous test.

ghstack-source-id: 44da621f749fff4a7dea247355da4de9e9a0ca73
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68183